Calculate next path point using lookahead for Graceful Controller (#5003)
* Add controller utils


* Fix linting issues


* Update regulated pure pursuit


* Update graceful controller


* Remove interpolate after goal & Use orientationAroundZAxis


* Update nav2_util/src/controller_utils.cpp


* Update nav2_util/src/controller_utils.cpp


* Update nav2_util/src/controller_utils.cpp


* Remove interpolate_after_goal parameter from test


---------
