+++ nav2_regulated_pure_pursuit_controller/test/test_regulated_pp.cpp
@@ -52,28 +52,6 @@ public:
     return getLookAheadDistance(twist);
   }
 
-  static geometry_msgs::msg::Point circleSegmentIntersectionWrapper(
-    const geometry_msgs::msg::Point & p1,
-    const geometry_msgs::msg::Point & p2,
-    double r)
-  {
-    return circleSegmentIntersection(p1, p2, r);
-  }
-
-  geometry_msgs::msg::PoseStamped
-  projectCarrotPastGoalWrapper(
-    const double & dist,
-    const nav_msgs::msg::Path & path)
-  {
-    return getLookAheadPoint(dist, path, true);
-  }
-
-  geometry_msgs::msg::PoseStamped getLookAheadPointWrapper(
-    const double & dist, const nav_msgs::msg::Path & path)
-  {
-    return getLookAheadPoint(dist, path);
-  }
-
   bool shouldRotateToPathWrapper(
     const geometry_msgs::msg::PoseStamped & carrot_pose, double & angle_to_path)
   {
@@ -207,239 +185,6 @@ TEST(RegulatedPurePursuitTest, findVelocitySignChange)
   EXPECT_EQ(rtn, std::numeric_limits<double>::max());
 }
 
-using CircleSegmentIntersectionParam = std::tuple<
-  std::pair<double, double>,
-  std::pair<double, double>,
-  double,
-  std::pair<double, double>
->;
-
-class CircleSegmentIntersectionTest
-  : public ::testing::TestWithParam<CircleSegmentIntersectionParam>
-{};
-
-TEST_P(CircleSegmentIntersectionTest, circleSegmentIntersection)
-{
-  auto pair1 = std::get<0>(GetParam());
-  auto pair2 = std::get<1>(GetParam());
-  auto r = std::get<2>(GetParam());
-  auto expected_pair = std::get<3>(GetParam());
-  auto pair_to_point = [](std::pair<double, double> p) -> geometry_msgs::msg::Point {
-      geometry_msgs::msg::Point point;
-      point.x = p.first;
-      point.y = p.second;
-      point.z = 0.0;
-      return point;
-    };
-  auto p1 = pair_to_point(pair1);
-  auto p2 = pair_to_point(pair2);
-  auto actual = BasicAPIRPP::circleSegmentIntersectionWrapper(p1, p2, r);
-  auto expected_point = pair_to_point(expected_pair);
-  EXPECT_DOUBLE_EQ(actual.x, expected_point.x);
-  EXPECT_DOUBLE_EQ(actual.y, expected_point.y);
-  // Expect that the intersection point is actually r away from the origin
-  EXPECT_DOUBLE_EQ(r, std::hypot(actual.x, actual.y));
-}
-
-INSTANTIATE_TEST_SUITE_P(
-  InterpolationTest,
-  CircleSegmentIntersectionTest,
-  testing::Values(
-    // Origin to the positive X axis
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {2.0, 0.0},
-  1.0,
-  {1.0, 0.0}
-},
-    // Origin to the negative X axis
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {-2.0, 0.0},
-  1.0,
-  {-1.0, 0.0}
-},
-    // Origin to the positive Y axis
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {0.0, 2.0},
-  1.0,
-  {0.0, 1.0}
-},
-    // Origin to the negative Y axis
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {0.0, -2.0},
-  1.0,
-  {0.0, -1.0}
-},
-    // non-origin to the X axis with non-unit circle, with the second point inside
-    CircleSegmentIntersectionParam{
-  {4.0, 0.0},
-  {-1.0, 0.0},
-  2.0,
-  {2.0, 0.0}
-},
-    // non-origin to the Y axis with non-unit circle, with the second point inside
-    CircleSegmentIntersectionParam{
-  {0.0, 4.0},
-  {0.0, -0.5},
-  2.0,
-  {0.0, 2.0}
-},
-    // origin to the positive X axis, on the circle
-    CircleSegmentIntersectionParam{
-  {2.0, 0.0},
-  {0.0, 0.0},
-  2.0,
-  {2.0, 0.0}
-},
-    // origin to the positive Y axis, on the circle
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {0.0, 2.0},
-  2.0,
-  {0.0, 2.0}
-},
-    // origin to the upper-right quadrant (3-4-5 triangle)
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {6.0, 8.0},
-  5.0,
-  {3.0, 4.0}
-},
-    // origin to the lower-left quadrant (3-4-5 triangle)
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {-6.0, -8.0},
-  5.0,
-  {-3.0, -4.0}
-},
-    // origin to the upper-left quadrant (3-4-5 triangle)
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {-6.0, 8.0},
-  5.0,
-  {-3.0, 4.0}
-},
-    // origin to the lower-right quadrant (3-4-5 triangle)
-    CircleSegmentIntersectionParam{
-  {0.0, 0.0},
-  {6.0, -8.0},
-  5.0,
-  {3.0, -4.0}
-}
-));
-
-TEST(RegulatedPurePursuitTest, projectCarrotPastGoal) {
-  auto ctrl = std::make_shared<BasicAPIRPP>();
-  auto node = std::make_shared<nav2::LifecycleNode>("testRPP");
-  std::string name = "PathFollower";
-  auto tf = std::make_shared<tf2_ros::Buffer>(node->get_clock());
-  auto costmap =
-    std::make_shared<nav2_costmap_2d::Costmap2DROS>("fake_costmap");
-  rclcpp_lifecycle::State state;
-  costmap->on_configure(state);
-  ctrl->configure(node, name, tf, costmap);
-
-  double EPSILON = std::numeric_limits<float>::epsilon();
-
-  nav_msgs::msg::Path path;
-  // More than 2 poses
-  path.poses.resize(4);
-  path.poses[0].pose.position.x = 0.0;
-  path.poses[1].pose.position.x = 1.0;
-  path.poses[2].pose.position.x = 2.0;
-  path.poses[3].pose.position.x = 3.0;
-  auto pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, 0.0, EPSILON);
-
-  // 2 poses fwd
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = 2.0;
-  path.poses[1].pose.position.x = 3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, 0.0, EPSILON);
-
-  // 2 poses at 45°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = 2.0;
-  path.poses[0].pose.position.y = 2.0;
-  path.poses[1].pose.position.x = 3.0;
-  path.poses[1].pose.position.y = 3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(45.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(45.0 * M_PI / 180) * 10.0, EPSILON);
-
-  // 2 poses at 90°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = 0.0;
-  path.poses[0].pose.position.y = 2.0;
-  path.poses[1].pose.position.x = 0.0;
-  path.poses[1].pose.position.y = 3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(90.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(90.0 * M_PI / 180) * 10.0, EPSILON);
-
-  // 2 poses at 135°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = -2.0;
-  path.poses[0].pose.position.y = 2.0;
-  path.poses[1].pose.position.x = -3.0;
-  path.poses[1].pose.position.y = 3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(135.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(135.0 * M_PI / 180) * 10.0, EPSILON);
-
-  // 2 poses back
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = -2.0;
-  path.poses[1].pose.position.x = -3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, -10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, 0.0, EPSILON);
-
-  // 2 poses at -135°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = -2.0;
-  path.poses[0].pose.position.y = -2.0;
-  path.poses[1].pose.position.x = -3.0;
-  path.poses[1].pose.position.y = -3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(-135.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(-135.0 * M_PI / 180) * 10.0, EPSILON);
-
-  // 2 poses at -90°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = 0.0;
-  path.poses[0].pose.position.y = -2.0;
-  path.poses[1].pose.position.x = 0.0;
-  path.poses[1].pose.position.y = -3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(-90.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(-90.0 * M_PI / 180) * 10.0, EPSILON);
-
-  // 2 poses at -45°
-  path.poses.clear();
-  path.poses.resize(2);
-  path.poses[0].pose.position.x = 2.0;
-  path.poses[0].pose.position.y = -2.0;
-  path.poses[1].pose.position.x = 3.0;
-  path.poses[1].pose.position.y = -3.0;
-  pt = ctrl->projectCarrotPastGoalWrapper(10.0, path);
-  EXPECT_NEAR(pt.pose.position.x, cos(-45.0 * M_PI / 180) * 10.0, EPSILON);
-  EXPECT_NEAR(pt.pose.position.y, sin(-45.0 * M_PI / 180) * 10.0, EPSILON);
-}
-
 TEST(RegulatedPurePursuitTest, lookaheadAPI)
 {
   auto ctrl = std::make_shared<BasicAPIRPP>();
@@ -476,24 +221,6 @@ TEST(RegulatedPurePursuitTest, lookaheadAPI)
   twist.linear.x = 0.0;
   rtn = ctrl->getLookAheadDistanceWrapper(twist);
   EXPECT_EQ(rtn, 0.3);
-
-  // test getLookAheadPoint
-  double dist = 1.0;
-  nav_msgs::msg::Path path;
-  path.poses.resize(10);
-  for (uint i = 0; i != path.poses.size(); i++) {
-    path.poses[i].pose.position.x = static_cast<double>(i);
-  }
-
-  // test exact hits
-  auto pt = ctrl->getLookAheadPointWrapper(dist, path);
-  EXPECT_EQ(pt.pose.position.x, 1.0);
-
-  // test interpolation
-  ctrl->configure(node, name, tf, costmap);
-  dist = 3.8;
-  pt = ctrl->getLookAheadPointWrapper(dist, path);
-  EXPECT_EQ(pt.pose.position.x, 3.8);
 }
 
 TEST(RegulatedPurePursuitTest, rotateTests)
